fix: Infinite loop in class loader chain causes SOE
Fixes: #2303
AtlasLibraryLoader is somehow set as a TCCL causes infinite loop
